Made wallet connect more modular
