+++ client/src/main.tsx
@@ -4,14 +4,17 @@ import './index.css'
 import App from './App.tsx'
 import { BrowserRouter as Router } from 'react-router-dom';
 import { AgentProvider } from './context/AgentContext';
+import AppWalletProvider from './components/WalletProvider';
 
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <AgentProvider>
-      <Router>
-        <App />
-      </Router>
+      <AppWalletProvider>
+        <Router>
+          <App />
+        </Router>
+      </AppWalletProvider>
     </AgentProvider>
   </StrictMode>,
 )
