+++ client/src/pages/TokenLaunchForm.tsx
@@ -1,9 +1,6 @@
-import React, { useState, useMemo, useEffect } from 'react';
-import { ConnectionProvider, WalletProvider, useWallet } from '@solana/wallet-adapter-react';
-import { WalletAdapterNetwork } from '@solana/wallet-adapter-base';
-import { PhantomWalletAdapter } from '@solana/wallet-adapter-wallets';
-import { WalletModalProvider } from '@solana/wallet-adapter-react-ui';
-import { clusterApiUrl, Connection, Keypair, LAMPORTS_PER_SOL, PublicKey } from '@solana/web3.js';
+import React, { useState,  useEffect } from 'react';
+import { useWallet } from '@solana/wallet-adapter-react';
+import { Connection, Keypair, LAMPORTS_PER_SOL } from '@solana/web3.js';
 import '@solana/wallet-adapter-react-ui/styles.css';
 import { createToken } from '../api/tokenAPI';
 import { encryptData } from '../utils/encryptionUtils';
@@ -14,9 +11,9 @@ import WalletConnectButton from '../components/WalletConnectButton';
 import AgentSelection from '../components/AgentSelection';
 import useCharacters from '../hooks/useCharacters'; // Import useCharacters
 import { useAgent } from '../context/AgentContext'; // Import the useAgent hook
-import { CheckCircle } from 'lucide-react'; // Import the checkmark icon
 import { fetchAraiBalance } from '../utils/fetchAraiBalance';
 import { addWalletRow, removeWalletRow } from '../TokenLaunchUtils/walletRowUtils';
+import ImagePreview from '../components/tokenCreationComponents/ImagePreview';
 
 
 
@@ -106,7 +103,7 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
         const mintAddress = response.data?.mintAddress || 'N/A';
         const url = response.data?.url || 'N/A';
         const txSignature = response.data?.transaction?.signature || 'N/A';
-        
+
         setTokenInfo({ mintAddress, url, txSignature });
 
         alert(
@@ -280,20 +277,20 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
   const handleRemoveWalletRow = (id: string) => removeWalletRow(id, setFormData);
 
   return (
-    <form 
-      onSubmit={handleSubmit} 
+    <form
+      onSubmit={handleSubmit}
       onChange={handleFormChange}
       className="w-full p-6 overflow-visible"
     >
-      <AgentSelection 
-        selectedCharacterIndex={selectedAgentIndex} 
-        handleSelectAgent={handleSelectAgent} 
+      <AgentSelection
+        selectedCharacterIndex={selectedAgentIndex}
+        handleSelectAgent={handleSelectAgent}
         className="mb-4 px-4 py-2"
       />
 
-      <WalletConnectButton 
-        araiTokenBalance={araiTokenBalance} 
-        formatNumberWithCommas={formatNumberWithCommas} 
+      <WalletConnectButton
+        araiTokenBalance={araiTokenBalance}
+        formatNumberWithCommas={formatNumberWithCommas}
       />
 
       <div className="max-w-4xl mx-auto space-y-6">
@@ -305,8 +302,8 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
 
         {/* Description */}
         <p className="text-gray-400 text-lg leading-relaxed">
-          During the Pump.fun launch, other addresses simultaneously perform token buy-in operations, 
-          effectively simplifying the trading process and accelerating market participation, 
+          During the Pump.fun launch, other addresses simultaneously perform token buy-in operations,
+          effectively simplifying the trading process and accelerating market participation,
           giving you an early advantage and potential profits sooner.
         </p>
 
@@ -368,68 +365,24 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
               Token image
               <span className="text-red-500 ml-1">*</span>
             </label>
-            <div className="flex gap-4 justify-center">
-              {/* Agent Image Preview */}
-              {agentImage && (
-                <div 
-                  className={`w-32 h-32 border-2 border-dashed border-orange-500/30 rounded-lg 
-                              flex items-center justify-center mb-4 relative cursor-pointer`}
-                  onClick={() => handleImageSelect(agentImage)}
-                >
-                  <img 
-                    alt="Agent image Preview"
-                    className="absolute inset-0 w-full h-full object-cover rounded-lg"
-                    src={agentImage}
-                  />
-                  {selectedImage === 'agent' && (
-                    <div className="absolute inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50">
-                      <CheckCircle className="w-8 h-8 text-white" />
-                    </div>
-                  )}
-                </div>
-              )}
-
-              {/* User Uploaded Image Preview */}
-              <div 
-                className="w-32 h-32 border-2 border-dashed border-orange-500/30 rounded-lg 
-                           flex items-center justify-center mb-4 relative cursor-pointer"
-                onClick={() => document.getElementById('image-upload')?.click()}
-              >
-                {formData.image instanceof File && (
-                  <img 
-                    id="image-preview"
-                    alt="Token image Preview"
-                    className="absolute inset-0 w-full h-full object-cover rounded-lg"
-                    src={URL.createObjectURL(formData.image)}
-                    style={{ display: 'block' }}
-                  />
-                )}
-                {selectedImage === 'uploaded' && (
-                  <div className="absolute inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50">
-                    <CheckCircle className="w-8 h-8 text-white" />
-                  </div>
-                )}
-                <span className="text-4xl text-orange-400">+</span>
-              </div>
-            </div>
+            <ImagePreview
+              agentImage={agentImage}
+              formDataImage={formData.image}
+              selectedImage={selectedImage}
+              handleImageSelect={handleImageSelect}
+              handleImageUpload={handleImageUpload}
+            />
             <p className="text-gray-100 mb-1">Select Agent Image or Click to Upload</p>
             <p className="text-gray-400">Supported formats: PNG/GIF/JPG/WEBP and JPEG</p>
             <p className="text-gray-400">Recommended size: 1000×1000 pixels</p>
             <p className="text-gray-400 mt-4 text-center">
               If it meets the above requirements, it can be better displayed on various platforms and applications.
             </p>
-            <input
-              id="image-upload"
-              type="file"
-              className="hidden"
-              accept="image/png,image/jpeg,image/jpg,image/gif,image/webp"
-              onChange={handleImageUpload}
-            />
           </div>
         </div>
 
         <div>
-        <label className="block text-lg text-gray-100 mb-2 pt-4">
+          <label className="block text-lg text-gray-100 mb-2 pt-4">
             Enter your Developer Wallet Info
           </label>
           <div className="overflow-x-auto">
@@ -437,7 +390,7 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
               <thead>
                 <tr>
                   <th className="text-left p-2 border-b border-orange-500/30">
-                    <span className="text-red-500">*</span> 
+                    <span className="text-red-500">*</span>
                     <span className="text-gray-100">Private Key</span>
                   </th>
                   <th className="text-left p-2 border-b border-orange-500/30">
@@ -468,13 +421,13 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
                     canRemoveRow={false}
                     showButtons={false}
                   />
-           )}
+                )}
               </tbody>
             </table>
           </div>
 
 
-          
+
 
         </div>
 
@@ -496,7 +449,7 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
                   onChange={handleChange}
                 />
               </div>
-              
+
               <div>
                 <label className="text-lg text-gray-100 mb-2 block">
                   Telegram:
@@ -530,23 +483,23 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
           </div>
         )}
 
-           {/* Wallet Parameters Section */}
-           <div className="border-t border-orange-500/20 pt-6">
+        {/* Wallet Parameters Section */}
+        <div className="border-t border-orange-500/20 pt-6">
           <div className="flex justify-between items-center mb-4">
             <h3 className="text-lg font-semibold text-gray-100">Enter Additional Jito Bundle Wallets</h3>
             <button className="text-orange-500 hover:text-orange-600 flex items-center gap-1">
               <span>↗</span> Batch Import Private Keys
             </button>
           </div>
-          
+
           <p className="text-gray-400 mb-4">Supports up to 20 wallet addresses for purchases. Over 16 addresses require two signatures.</p>
-          
+
           <div className="overflow-x-auto">
             <table className="w-full border-collapse">
               <thead>
                 <tr>
                   <th className="text-left p-2 border-b border-orange-500/30">
-                    <span className="text-red-500">*</span> 
+                    <span className="text-red-500">*</span>
                     <span className="text-gray-100">Private Key</span>
                   </th>
                   <th className="text-left p-2 border-b border-orange-500/30">
@@ -669,13 +622,12 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
 
 
         {/* Submit Button */}
-        <button 
+        <button
           type="submit"
-          className={`w-full py-3 rounded-md transition-colors ${
-            araiTokenBalance !== null && araiTokenBalance >= 100000
+          className={`w-full py-3 rounded-md transition-colors ${araiTokenBalance !== null && araiTokenBalance >= 100000
               ? 'bg-gradient-to-r from-cyan-600 to-orange-600 hover:from-cyan-700 hover:to-orange-700 text-white'
               : 'bg-gray-500 text-gray-300 cursor-not-allowed'
-          }`}
+            }`}
           disabled={araiTokenBalance === null || araiTokenBalance < 100000}
         >
           Launch Pumpfun Token
@@ -696,40 +648,4 @@ const TokenLaunchForm: React.FC<TokenLaunchFormProps> = ({ formData, setFormData
   );
 };
 
-// Create a wrapper component to provide wallet context
-const TokenLaunchFormWithWallet: React.FC<TokenLaunchFormProps> = (props) => {
-  const network = WalletAdapterNetwork.Mainnet;
-  const endpoint = useMemo(() => {
-    // Use the HELIUS_RPC_URL from environment variables
-    const url = import.meta.env.VITE_HELIUS_RPC_URL || clusterApiUrl(network);
-    console.log('Using RPC endpoint:', url);
-    
-    // Validate endpoint URL
-    if (!url.startsWith('https://')) {
-      console.error('Invalid RPC endpoint:', url);
-      return clusterApiUrl(network); // Fallback to default
-    }
-    
-    console.log('Using RPC endpoint:', url);
-    return url;
-  }, [network]);
-  
-  const wallets = useMemo(
-    () => [
-      new PhantomWalletAdapter(),
-    ],
-    []
-  );
-
-  return (
-    <ConnectionProvider endpoint={endpoint}>
-      <WalletProvider wallets={wallets} autoConnect>
-        <WalletModalProvider>
-          <TokenLaunchForm {...props} />
-        </WalletModalProvider>
-      </WalletProvider>
-    </ConnectionProvider>
-  );
-};
-
-export default TokenLaunchFormWithWallet;
+export default TokenLaunchForm;
